Growth 4

+++ server/internal/models/model.go
@@ -36,7 +36,7 @@ type EnergyRecord struct {
 type GrowthRecord struct {
 	ID               uint    `gorm:"primaryKey" json:"id"`
 	Date             string  `gorm:"not null;index" json:"date"`              // "2026-05-27"
-	LeafCount        int     `json:"leaf_count"`                              // จำนวนใบที่ตรวจพบ
+	LeafCount        float64 `json:"leaf_count"`                              // จำนวนใบเฉลี่ยต่อต้น
 	PlantCount       int     `json:"plant_count"`                             // จำนวนต้นที่ตรวจพบ
 	HarvestReadiness float64 `json:"harvest_readiness"`                       // % ความพร้อมเก็บเกี่ยว
 	ImageURL         string  `json:"image_url,omitempty"`                     // URL รูปที่สแกน
